--- tomcat/src/main/java/org/apache/coyote/http11/Controller.java
@@ -1,6 +1,0 @@
-package org.apache.coyote.http11;
-
-public interface Controller {
-
-    HttpResponse process(String uri);
-}
