+++ tomcat/src/main/java/org/apache/coyote/http11/controller/Controller.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.http11.controller;
+
+import org.apache.coyote.http11.httpresponse.HttpResponse;
+
+public interface Controller {
+
+    HttpResponse process(String uri);
+}
