+++ tomcat/src/main/java/org/apache/coyote/http11/httprequest/Body.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httprequest;
 
 public class Body {
