+++ tomcat/src/main/java/org/apache/coyote/http11/httprequest/Headers.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httprequest;
 
 import java.util.LinkedHashMap;
 import java.util.List;
