+++ tomcat/src/test/java/org/apache/coyote/http11/httprequest/HttpRequestTest.java
@@ -1,5 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httprequest;
 
+import org.apache.coyote.http11.InputReader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
